requirements.txt
pymongo
flask
flask_cors
python-dotenv
gunicorn==20.1.0